+++ CHANGELOG.md
@@ -16,8 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New payment's module architecture (#2135)
 - Established Vuex naming conventions. TLDR - we strongly recommend to use vuex getters instead of mapping state itself (#2069)
 - Fixed isServer flag in module hooks (#840)
-- `store/lib/search` has been moved to `core/lib/search`. All search related types from `store/types/search` are now exported in `core/lib/search.ts` (#2225)
+- Location of type files (#2226)
+- `store/lib/search` has been moved to `core/lib/search` (#2225)
 - `store/lib/multistore` has been moved to `core/lib/multistore` (#2224)
+
 
 ## [1.6.0] - 2018.12.05